+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Changelog
 This is a high-level changelog for each released versions of the plugin.
 For a more detailed list of past and incoming changes, see the commit history.
 
+1.0.x
+------
+- Added support to delete messages. DiscordBot.delete()
+- Fixed bot crashing when only stickers were sent
+- Made Helpers.make_iso_string() more efficient
+- Fixed typo parsing options in User.get_display_avatar
+
 1.0.0
 ------